# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ActivityListResponseDto": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ActivityResponseDto"
            },
            "type": "array"
          },
          "nextCursor": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
      "ActivityResponseDto": {
        "properties": {
          "actor": {
            "allOf": [
              {
                "$ref": "#/components/schemas/CommentAuthorDto"
              }
            ],
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "eventType": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "subjectId": {
            "format": "uuid",
            "type": "string"
          },
          "subjectType": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "eventType",
          "subjectType",
          "subjectId",
          "metadata",
          "actor",
          "createdAt"
        ],
        "type": "object"
      },
      "ArchiveDocumentRequestDto": {
        "properties": {
          "contentVersion": {
            "type": "number"
          }
        },
        "required": [
          "contentVersion"
        ],
        "type": "object"
      },
...
      "CommentAuthorDto": {
        "properties": {
          "avatarUrl": {
            "nullable": true,
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "initials": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "displayName",
          "initials",
          "avatarUrl"
        ],
        "type": "object"
      },
      "CommentListResponseDto": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CommentResponseDto"
            },
            "type": "array"
          },
          "nextCursor": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "items"
        ],
        "type": "object"
      },
      "CommentRequestDto": {
        "properties": {
          "body": {
            "maxLength": 10000,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "body"
        ],
        "type": "object"
      },
      "CommentResponseDto": {
        "properties": {
          "author": {
            "$ref": "#/components/schemas/CommentAuthorDto"
          },
          "body": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "taskId": {
            "format": "uuid",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "taskId",
          "body",
          "author",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "CreateDocumentRequestDto": {
        "properties": {
          "content": {
            "type": "string"
          },
          "projectId": {
            "nullable": true,
            "type": "string"
          },
          "sectionId": {
            "nullable": true,
            "type": "string"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "type": "object"
      },
...
      "DocumentResponseDto": {
        "properties": {
          "archivedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "contentVersion": {
            "type": "number"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "projectId": {
            "nullable": true,
            "type": "string"
          },
          "sectionId": {
            "nullable": true,
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "workspaceId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "workspaceId",
          "title",
          "content",
          "contentVersion",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
...
      "MemberSummaryDto": {
        "properties": {
          "avatarUrl": {
            "nullable": true,
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "initials": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "displayName",
          "initials",
          "avatarUrl"
        ],
        "type": "object"
      },
...
      "TaskResponseDto": {
        "properties": {
          "archivedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "assignee": {
            "allOf": [
              {
                "$ref": "#/components/schemas/MemberSummaryDto"
              }
            ],
            "nullable": true
          },
          "assigneeId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "completedAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "description": {
            "nullable": true,
            "type": "string"
          },
          "dueAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "parentTaskId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "position": {
            "type": "number"
          },
          "priority": {
            "enum": [
              "URGENT",
              "HIGH",
              "NORMAL",
              "LOW"
            ],
            "type": "string"
          },
          "projectId": {
            "format": "uuid",
            "type": "string"
          },
          "sectionId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "statusId": {
            "format": "uuid",
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          },
          "version": {
            "type": "number"
          },
          "workspaceId": {
            "format": "uuid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "workspaceId",
          "projectId",
          "sectionId",
          "statusId",
          "assigneeId",
          "parentTaskId",
          "title",
          "description",
          "priority",
          "position",
          "dueAt",
          "completedAt",
          "version",
          "createdAt",
          "updatedAt",
          "archivedAt",
          "assignee"
        ],
        "type": "object"
      },
...
      "UpdateDocumentRequestDto": {
        "properties": {
          "content": {
            "type": "string"
          },
          "contentVersion": {
            "type": "number"
          },
          "title": {
            "type": "string"
          }
        },
        "required": [
          "contentVersion"
        ],
        "type": "object"
      },
...
      "UpdateTaskRequestDto": {
        "properties": {
          "assigneeId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "description": {
            "maxLength": 20000,
            "nullable": true,
            "type": "string"
          },
          "dueAt": {
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "parentTaskId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "priority": {
            "enum": [
              "URGENT",
              "HIGH",
              "NORMAL",
              "LOW"
            ],
            "type": "string"
          },
          "sectionId": {
            "format": "uuid",
            "nullable": true,
            "type": "string"
          },
          "statusId": {
            "format": "uuid",
            "type": "string"
          },
          "title": {
            "maxLength": 240,
            "type": "string"
          },
          "version": {
            "type": "number"
          }
        },
        "required": [
          "version"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Versioned REST API for ClickFlow workspaces and projects.",
    "title": "ClickFlow API",
    "version": "1.0.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/workspaces/{workspaceId}/documents": {
      "get": {
        "operationId": "DocumentController_list",
        "parameters": [
          {
            "in": "query",
            "name": "includeArchived",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "projectId",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DocumentResponseDto"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      },
      "post": {
        "operationId": "DocumentController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/documents/{documentId}": {
      "delete": {
        "operationId": "DocumentController_archive",
        "parameters": [
          {
            "in": "path",
            "name": "documentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArchiveDocumentRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      },
      "get": {
        "operationId": "DocumentController_get",
        "parameters": [
          {
            "in": "path",
            "name": "documentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      },
      "patch": {
        "operationId": "DocumentController_update",
        "parameters": [
          {
            "in": "path",
            "name": "documentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "documents"
        ]
      }
    },
...
    "/api/v1/workspaces/{workspaceId}/tasks/{taskId}": {
      "delete": {
        "operationId": "TaskController_archive",
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "version",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tasks"
        ]
      },
      "get": {
        "operationId": "TaskController_get",
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tasks"
        ]
      },
      "patch": {
        "operationId": "TaskController_update",
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/tasks/{taskId}/activity": {
      "get": {
        "operationId": "CommentController_activity",
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActivityListResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "comments",
          "activity"
        ]
      }
    },
...
    "/api/v1/workspaces/{workspaceId}/tasks/{taskId}/comments": {
      "get": {
        "operationId": "CommentController_list",
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentListResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "comments",
          "activity"
        ]
      },
      "post": {
        "operationId": "CommentController_create",
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "comments",
          "activity"
        ]
      }
    },
    "/api/v1/workspaces/{workspaceId}/tasks/{taskId}/comments/{commentId}": {
      "delete": {
        "operationId": "CommentController_remove",
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "commentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "comments",
          "activity"
        ]
      },
      "patch": {
        "operationId": "CommentController_update",
        "parameters": [
          {
            "in": "path",
            "name": "taskId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "commentId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommentRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommentResponseDto"
                }
              }
            },
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "comments",
          "activity"
        ]
      }
    },
...
  },
  "servers": [],
...
}